+++ dbt-ml/src/dbt_ml/checks/__init__.py
@@ -1,5 +1,11 @@
 from ..test_specs import SUPPORTED_TESTS
-from .runner import run_model_tests, run_project_tests
+from .runner import run_model_tests, run_project_tests, validate_test_requirements
 from .schema import TestResult
 
-__all__ = ["SUPPORTED_TESTS", "TestResult", "run_model_tests", "run_project_tests"]
+__all__ = [
+    "SUPPORTED_TESTS",
+    "TestResult",
+    "run_model_tests",
+    "run_project_tests",
+    "validate_test_requirements",
+]